Add files via upload
For golfsim.cpp: Added the ability for the user to specify the name of the file that holds the golfer names and ratings. Also added some error checking related to that, and the attribute of nationality to go with my future plans to implement a Ryder Cup mode to the simulator.
For golferratings.txt: Added characters to be read into the nationality attribute for each golfer as preparation for my Ryder Cup mode plans.