FedEx Bug Fix
Just added a for loop that makes everyone's FedEx points 0 before the simulation starts. I didn't think that would be necessary, but it seems to have fixed the bug since I haven't been able to make it happen again. I suppose there's still a slim possibility the bug is still there, but if anyone ever reads this, by the time they do I'm sure it'll be gone.