Players Mode, True Round Scores: Big March Change
It's been a while since I've changed this code, huh? But here it is, my big March update. The biggest things are the addition of the Players 2020 mode (RIP) and the generation of a true stroke play score for each player in Career and Players mode using randomly generated numbers with a mean and SD. Also, in these modes the final results are printed in a table instead of just in rows.